+++ test/support/rails_app/config/initializers/devise_async.rb
@@ -1,0 +1,3 @@
+Devise::Async.setup do |config|
+  config.queue = :custom_queue
+end