+++ test/test_helper.rb
@@ -11,6 +11,7 @@ require "devise"
 require "resque"
 require "sidekiq"
 require "delayed_job_active_record"
+require "sidekiq/testing"
 
 require "support/rails_app"
 require "support/helpers"